requirements.txt
python-dotenv==1.0.0
playwright==1.40.0
openai==1.6.1
python-telegram-bot==20.7
asyncio==3.4.3
aiohttp==3.9.1
python-dateutil==2.8.2 